# src/server/swagger.yaml
openapi: 3.0.0
info:
  title: Kroombox API Gateway
  description: Dokumentasi Lengkap API Gateway untuk Kroombox, mencakup manajemen klien, proxy, metrik GPU, event realtime, dan panel admin.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
  - url: "{appUrl}"
    description: Production server
    variables:
      appUrl:
        default: https://api.kroombox.com
        description: Base URL untuk Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Gunakan format: Bearer {API_KEY}'
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /gateway/{target}:
    get:
      tags: [Proxy Gateway]
      summary: Forward Request ke Upstream Target (GET)
      parameters:
        - in: path
          name: target
          required: true
          schema:
            type: string
          description: Parameter Route Target (misal whatsapp)
      responses:
        '200':
          description: Proxy Response
    post:
      tags: [Proxy Gateway]
      summary: Forward Request ke Upstream Target (POST)
      parameters:
        - in: path
          name: target
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Proxy Response

  # ---------------------------------------------------------
  # EVENT BUS
  # ---------------------------------------------------------
...